data-crawler/requirements.txt
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
boto3==1.33.6
pymongo==4.6.0
fastapi==0.104.1
pydantic==2.5.2
celery==5.3.6
redis==5.0.1
playwright==1.40.0 